[MS-1229] move helpers funcs into helpers folder
