+++ src/atomic/organism/review.tsx
@@ -10,6 +10,7 @@ const LG_BOOTSTRAP = 992;
 export default function Review() {
     const reviews = useReviewData();
     const isMobile = useMediaQuery(`(max-width: ${LG_BOOTSTRAP}px)`);
+    console.log(reviews);
 
     if (isMobile) {
         return <ReviewMobile reviews={reviews} />;
